# lib/api-zod/src/generated/types/postInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PostInput {
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  category: string;
  excerpt: string;
  content: string;
  imageUrl: string;
}